refactor: 각 클래스들을 용도에 맞게 package별로 나눔
