+++ src/main/java/baseball/Application.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import baseball.controller.PlayGroundManager;
+
 import nextstep.utils.Console;
 
 public class Application {
