+++ src/main/java/baseball/enums/MATCH_RESULT.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.enums;
 
 public enum MATCH_RESULT {
     NOTHING,
