+++ src/main/java/baseball/enums/REGAME_FLAG.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.enums;
 
 public enum REGAME_FLAG {
     NONE,
