+++ src/main/java/baseball/model/PlayGround.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.model;
 
 public class PlayGround {
     private PlayerState computerPlayerState;
