+++ src/main/java/baseball/utils/InputParser.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.utils;
 
 import java.util.ArrayList;
