+++ src/test/java/baseball/model/BallCollectionTest.java
@@ -1,9 +1,11 @@
-package baseball;
+package baseball.model;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import baseball.enums.MATCH_RESULT;
 
 public class BallCollectionTest {
     private BallCollection collection;
