+++ src/test/java/baseball/model/PlayGroundTest.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.model;
 
 import java.util.Arrays;
