+++ src/test/java/baseball/utils/InputParserTest.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.utils;
 
 import java.util.ArrayList;
